cyber_actuarial_risk_simulator/requirements.txt
Flask-JWT-Extended
Flask-Security-Too
Flask-Bcrypt
Flask-Limiter
Flask-SQLAlchemy
Flask-Migrate
Flask-Dance
python-dotenv 
flask-jwt-extended
pyotp
redis
pydantic
zxcvbn
cryptography
psycopg2-binary 